Move the controller into a separate file
